+++ src/@next/components/molecules/InputSelect/InputSelect.tsx
@@ -21,6 +21,7 @@ export const InputSelect: React.FC<IProps> = ({ label, ...props }: IProps) => {
         outlineStyle: "solid",
         outlineWidth: "1px",
       },
+      background: "none",
       border: state.menuIsOpen
         ? `1px solid ${secondaryColor}`
         : `1px solid ${borderColor}`,
@@ -38,6 +39,7 @@ export const InputSelect: React.FC<IProps> = ({ label, ...props }: IProps) => {
       };
     },
   };
+
   const customComponents = {
     Control: (props: any) => {
       const customTheme = React.useContext(ThemeContext);
@@ -46,7 +48,7 @@ export const InputSelect: React.FC<IProps> = ({ label, ...props }: IProps) => {
           <components.Control {...{ customTheme, ...props }} />
           {
             <InputLabel
-              labelBackground="#FFF"
+              labelBackground={customTheme.colors.light}
               active={props.selectProps.menuIsOpen || props.hasValue}
             >
               {label}
@@ -56,14 +58,26 @@ export const InputSelect: React.FC<IProps> = ({ label, ...props }: IProps) => {
       );
     },
     IndicatorSeparator: () => null,
-    IndicatorsContainer: ({ selectProps }: any) => {
-      return (
+    IndicatorsContainer: ({ selectProps, hasValue, clearValue }: any) => {
+      const showClearIndicator =
+        selectProps.isClearable ||
+        (selectProps.isMulti && selectProps.isClearable === undefined);
+
+      if (showClearIndicator && hasValue) {
+        return (
+          <S.ClearIndicator onClick={clearValue}>
+            <Icon name="select_x" size={10} />
+          </S.ClearIndicator>
+        );
+      } else {
         // Boolean to string conversion done due to
         // https://github.com/styled-components/styled-components/issues/1198
-        <S.Indicator rotate={String(selectProps.menuIsOpen)}>
-          <Icon name="select_arrow" size={10} />
-        </S.Indicator>
-      );
+        return (
+          <S.DropdownIndicator rotate={String(selectProps.menuIsOpen)}>
+            <Icon name="select_arrow" size={10} />
+          </S.DropdownIndicator>
+        );
+      }
     },
     Option: (props: any) => {
       const customTheme = React.useContext(ThemeContext);