+++ src/@next/components/organisms/ProductVariantPicker/styles.ts
@@ -1,0 +1,11 @@
+import { styled } from "@styles";
+
+export const Wrapper = styled.div`
+  display: grid;
+  grid-gap: 20px;
+  grid-template-columns: auto auto;
+`;
+export const SelectIndicator = styled.div`
+  margin: 0 1rem 0 0;
+  cursor: pointer;
+`;