+++ src/@next/components/organisms/ProductVariantPicker/types.ts
@@ -1,0 +1,13 @@
+import { ProductDetails_product_variants } from "@sdk/queries/types/ProductDetails";
+
+import { IProductVariantsAttributesSelectedValues } from "@types";
+
+export interface IProps {
+  productVariants?: ProductDetails_product_variants[];
+  onChange?: (
+    selectedAttributesValues?: IProductVariantsAttributesSelectedValues,
+    selectedVariant?: ProductDetails_product_variants | undefined
+  ) => void;
+  selectSidebar?: boolean;
+  selectSidebarTarget?: HTMLElement | null;
+}