+++ src/@next/components/organisms/SizeOverlay/types.ts
@@ -1,8 +1,0 @@
-export interface IProps {
-  hide: () => void;
-  onClick: (value: string) => void;
-  selected?: string;
-  show: boolean;
-  target?: HTMLElement | null;
-  values: string[];
-}