+++ src/@next/types/ISelectOptions.ts
@@ -1,0 +1,4 @@
+export interface ISelectOption {
+  label: string;
+  value: string;
+}