+++ src/@sdk/mutations/types/SetCustomerDefaultAddress.ts
@@ -2,7 +2,7 @@
 /* eslint-disable */
 // This file was automatically generated and should not be edited.
 
-import { AddressTypeEnum } from "./../../../../types/globalTypes";
+import { AddressTypeEnum } from "./../../types/globalTypes";
 
 // ====================================================
 // GraphQL mutation operation: SetCustomerDefaultAddress
@@ -52,6 +52,14 @@ export interface SetCustomerDefaultAddress_accountSetDefaultAddress_user_default
   country: SetCustomerDefaultAddress_accountSetDefaultAddress_user_defaultShippingAddress_country;
   countryArea: string;
   phone: string | null;
+  /**
+   * Address is user's default billing address.
+   */
+  isDefaultBillingAddress: boolean | null;
+  /**
+   * Address is user's default shipping address.
+   */
+  isDefaultShippingAddress: boolean | null;
 }
 
 export interface SetCustomerDefaultAddress_accountSetDefaultAddress_user_defaultBillingAddress_country {
@@ -85,6 +93,14 @@ export interface SetCustomerDefaultAddress_accountSetDefaultAddress_user_default
   country: SetCustomerDefaultAddress_accountSetDefaultAddress_user_defaultBillingAddress_country;
   countryArea: string;
   phone: string | null;
+  /**
+   * Address is user's default billing address.
+   */
+  isDefaultBillingAddress: boolean | null;
+  /**
+   * Address is user's default shipping address.
+   */
+  isDefaultShippingAddress: boolean | null;
 }
 
 export interface SetCustomerDefaultAddress_accountSetDefaultAddress_user_addresses_country {
@@ -118,6 +134,14 @@ export interface SetCustomerDefaultAddress_accountSetDefaultAddress_user_address
   country: SetCustomerDefaultAddress_accountSetDefaultAddress_user_addresses_country;
   countryArea: string;
   phone: string | null;
+  /**
+   * Address is user's default billing address.
+   */
+  isDefaultBillingAddress: boolean | null;
+  /**
+   * Address is user's default shipping address.
+   */
+  isDefaultShippingAddress: boolean | null;
 }
 
 export interface SetCustomerDefaultAddress_accountSetDefaultAddress_user {