+++ src/views/Product/types/ProductDetails.ts
@@ -547,8 +547,28 @@ export interface ProductDetails_product_images {
   url: string;
 }
 
+export interface ProductDetails_product_basePrice {
+  __typename: "Money";
+  /**
+   * Amount of money.
+   */
+  amount: number;
+  /**
+   * Currency code.
+   */
+  currency: string;
+  /**
+   * Money formatted according to the current locale.
+   */
+  localized: string;
+}
+
 export interface ProductDetails_product_attributes_attribute {
   __typename: "Attribute";
+  /**
+   * The ID of the object.
+   */
+  id: string;
   /**
    * Name of an attribute displayed in the interface.
    */
@@ -557,6 +577,10 @@ export interface ProductDetails_product_attributes_attribute {
 
 export interface ProductDetails_product_attributes_values {
   __typename: "AttributeValue";
+  /**
+   * The ID of the object.
+   */
+  id: string;
   /**
    * Name of a value displayed in the interface.
    */
@@ -778,6 +802,10 @@ export interface ProductDetails_product {
    * List of images for the product.
    */
   images: (ProductDetails_product_images | null)[] | null;
+  /**
+   * The product's default base price.
+   */
+  basePrice: ProductDetails_product_basePrice | null;
   /**
    * List of attributes assigned to this product.
    */