+++ types/globalTypes.ts
@@ -352,9 +352,9 @@ export interface PaymentInput {
 }
 
 export interface ProductOrder {
-  field?: ProductOrderField | null;
-  attributeId?: string | null;
   direction: OrderDirection;
+  attributeId?: string | null;
+  field?: ProductOrderField | null;
 }
 
 //==============================================================